requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
google-generativeai==0.4.0